+++ changelog.md
@@ -2,3 +2,4 @@
 - Improved Surface Detection (Goop now sticks to non-full block surfaces like slabs)
 - goop is no longer offset weirdly in negative coordinates
 - Added Random Rotation Setting
+- Added Keybind for clearing all Goop